+++ shuup/front/apps/carousel/admin_module/views/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+from .delete import CarouselDeleteView
+from .edit import CarouselEditView
+from .list import CarouselListView
+
+__all__ = [
+    "CarouselDeleteView",
+    "CarouselEditView",
+    "CarouselListView"
+]